move env tests
