+++ src/__tests__/config.test.ts
@@ -1,29 +1,14 @@
 /**
- * Tests of our ability to read craft config files and env vars (NOT general
- * test configuration).
+ * Tests of our ability to read craft config files. (This is NOT general test
+ * configuration).
  */
 
-import { readFileSync, writeFileSync } from 'fs';
-import { homedir } from 'os';
+import { readFileSync } from 'fs';
 import { dirname, join } from 'path';
 
 import { compile } from 'json-schema-to-typescript';
 
-import {
-  CONFIG_FILE_NAME,
-  ENV_FILE_NAME,
-  getProjectConfigSchema,
-  readEnvironmentConfig,
-  validateConfiguration,
-} from '../config';
-import { withTempDir } from '../utils/files';
-
-jest.mock('os', () => ({ ...require.requireActual('os'), homedir: jest.fn() }));
-
-// Mock getConfigFileDir
-// tslint:disable-next-line:no-var-requires
-const configModule = require('../config');
-const getConfigFileDirMock = (configModule.getConfigFileDir = jest.fn());
+import { getProjectConfigSchema, validateConfiguration } from '../config';
 
 const configSchemaDir = join(dirname(__dirname), 'schemas');
 const configGeneratedTypes = join(configSchemaDir, 'project_config.ts');
@@ -76,124 +61,5 @@ describe('getProjectConfigSchema', () => {
     expect(projectConfigSchema).toBeTruthy();
     expect(projectConfigSchema).toHaveProperty('title');
     expect(projectConfigSchema).toHaveProperty('properties');
-  });
-});
-
-describe('readEnvironmentConfig', () => {
-  // const getConfigFileDirMock = getConfigFileDir as jest.Mock;
-  const homedirMock = homedir as jest.Mock;
-
-  const invalidDir = '/invalid/invalid';
-
-  function writeConfigFile(directory: string): void {
-    const outConfigFile = join(directory, CONFIG_FILE_NAME);
-    writeFileSync(outConfigFile, '');
-  }
-
-  beforeEach(() => {
-    delete process.env.TEST_BLA;
-    jest.clearAllMocks();
-  });
-  test('calls homedir/findConfigFile', async () => {
-    process.env.TEST_BLA = '123';
-
-    homedirMock.mockReturnValue(invalidDir);
-    getConfigFileDirMock.mockReturnValue(invalidDir);
-
-    readEnvironmentConfig();
-
-    expect(getConfigFileDirMock).toHaveBeenCalledTimes(1);
-    expect(homedirMock).toHaveBeenCalledTimes(1);
-    expect(process.env.TEST_BLA).toBe('123');
-    expect(ENV_FILE_NAME.length).toBeGreaterThanOrEqual(1);
-  });
-
-  test('checks the config directory', async () => {
-    homedirMock.mockReturnValue(invalidDir);
-    delete process.env.TEST_BLA;
-
-    await withTempDir(async directory => {
-      getConfigFileDirMock.mockReturnValue(directory);
-
-      writeConfigFile(directory);
-
-      const outFile = join(directory, ENV_FILE_NAME);
-      writeFileSync(outFile, 'export TEST_BLA=234\nexport TEST_ANOTHER=345');
-
-      readEnvironmentConfig();
-
-      expect(process.env.TEST_BLA).toBe('234');
-    });
-  });
-  test('checks home directory', async () => {
-    getConfigFileDirMock.mockReturnValue(invalidDir);
-    delete process.env.TEST_BLA;
-
-    await withTempDir(async directory => {
-      homedirMock.mockReturnValue(directory);
-      const outFile = join(directory, ENV_FILE_NAME);
-      writeFileSync(outFile, 'export TEST_BLA=234\n');
-
-      readEnvironmentConfig();
-
-      expect(process.env.TEST_BLA).toBe('234');
-    });
-  });
-
-  test('checks home directory first, and then the config directory', async () => {
-    delete process.env.TEST_BLA;
-
-    await withTempDir(async dir1 => {
-      await withTempDir(async dir2 => {
-        homedirMock.mockReturnValue(dir1);
-
-        const outHome = join(dir1, ENV_FILE_NAME);
-        writeFileSync(outHome, 'export TEST_BLA=from_home');
-
-        getConfigFileDirMock.mockReturnValue(dir2);
-        writeConfigFile(dir2);
-        const configDirFile = join(dir2, ENV_FILE_NAME);
-        writeFileSync(configDirFile, 'export TEST_BLA=from_config_dir');
-
-        readEnvironmentConfig();
-
-        expect(process.env.TEST_BLA).toBe('from_config_dir');
-      });
-    });
-  });
-
-  test('does not overwrite existing variables by default', async () => {
-    homedirMock.mockReturnValue(invalidDir);
-
-    process.env.TEST_BLA = 'existing';
-
-    await withTempDir(async directory => {
-      getConfigFileDirMock.mockReturnValue(directory);
-      const outFile = join(directory, ENV_FILE_NAME);
-      writeFileSync(outFile, 'export TEST_BLA=new_value');
-
-      readEnvironmentConfig();
-
-      expect(process.env.TEST_BLA).toBe('existing');
-    });
-  });
-
-  test('overwrites existing variables if explicitly stated', async () => {
-    homedirMock.mockReturnValue(invalidDir);
-
-    process.env.TEST_BLA = 'existing';
-
-    await withTempDir(async directory => {
-      getConfigFileDirMock.mockReturnValue(directory);
-
-      writeConfigFile(directory);
-
-      const outFile = join(directory, ENV_FILE_NAME);
-      writeFileSync(outFile, 'export TEST_BLA=new_value');
-
-      readEnvironmentConfig(true);
-
-      expect(process.env.TEST_BLA).toBe('new_value');
-    });
   });
 });
